Use provider and class for Firestore operation, change usecase naming to api
