+++ flutter_src/lib/logic/state/team_state/team_bloc.dart
@@ -38,7 +38,7 @@ class TeamFormBloc extends FormBloc<String, String> {
     );
 
     //TODO
-    // Upload Team Object
+    //Upload Team Object
   }
 
   @override
