+++ flutter_src/lib/presentation/pages/macro_table_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:macrobaseapp/logic/api/firestore_db.dart';
 import 'package:macrobaseapp/presentation/widgets/misc/no_macro_illustration.dart';
 import 'package:macrobaseapp/presentation/widgets/component/macro_table_entry.dart';
 import 'package:macrobaseapp/logic/state/macro_state/macro_notifier.dart';
-import 'package:macrobaseapp/logic/usecases/macro_firestore/firestore_macro_operation.dart';
 import 'package:macrobaseapp/model/entities/user.dart';
 import 'package:provider/provider.dart';
 
@@ -16,7 +16,7 @@ class _MacroTableState extends State<MacroTable> {
   Widget build(BuildContext context) {
     final macroNotifier = Provider.of<MacroNotifier>(context, listen: true);
     final user = Provider.of<User>(context);
-    getMacros(macroNotifier, user.email);
+    context.read<FirestoreService>().getMacros(macroNotifier, user.email);
 
     if (macroNotifier.macroList.length == 0) {
       return NoMacroIllustration();
